# src/main/java/io/swagger/api/IdApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.41).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.model.Server;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.bind.annotation.CookieValue;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2023-03-25T01:54:50.682052577Z[GMT]")
@Validated
public interface IdApi {

    @Operation(summary = "Deletes a server", description = "delete a server by ID", tags={  })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "202", description = "server deleted successfully"),
        
        @ApiResponse(responseCode = "404", description = "server not found") })
    @RequestMapping(value = "/{id}",
        method = RequestMethod.DELETE)
    ResponseEntity<Void> deleteById(@Parameter(in = ParameterIn.PATH, description = "server with this id to delete", required=true, schema=@Schema()) @PathVariable("id") String id);


    @Operation(summary = "", description = "Returns a server based on ID", tags={  })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "server found successfully", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Server.class))),
        
        @ApiResponse(responseCode = "404", description = "server not found") })
    @RequestMapping(value = "/{id}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<Server> returnById(@Parameter(in = ParameterIn.PATH, description = "server ID", required=true, schema=@Schema()) @PathVariable("id") String id);

}