# src/main/java/io/swagger/api/NameApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.41).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.model.Server;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.bind.annotation.CookieValue;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2023-03-25T01:54:50.682052577Z[GMT]")
@Validated
public interface NameApi {

    @Operation(summary = "", description = "Returns servers based on this string in their name", tags={  })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "servers found successfully", content = @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Server.class)))),
        
        @ApiResponse(responseCode = "404", description = "no server found") })
    @RequestMapping(value = "/name/{name}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<List<Server>> returnByName(@Parameter(in = ParameterIn.PATH, description = "server names", required=true, schema=@Schema()) @PathVariable("name") String name);

}